+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-07-28
+
 ### Fixed
 
 - Explicitly allow reading the `fhirUser` resource when scope and launch context ID are present (RFS-309)
@@ -67,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump `micrometer.version` from 1.16.2 to 1.16.5 ([#12](https://github.com/OHSU-OCTRI/fhir-sandbox-backend/pull/12))
 
-[unreleased]: https://github.com/OHSU-OCTRI/fhir-sandbox-backend/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/OHSU-OCTRI/fhir-sandbox-backend/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/OHSU-OCTRI/fhir-sandbox-backend/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/OHSU-OCTRI/fhir-sandbox-backend/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/OHSU-OCTRI/fhir-sandbox-backend/compare/4660cfbad3ecba599db212826fab3392c150e6bc...v0.1.0